feat(fluxion-fluid): define HvacState and HvacMode types
Moves HvacState and HvacMode from fluxion-grid to fluxion-fluid as
part of the crate split. These types are now centralized in fluxion-fluid
for use across the workspace.

Closes #2274
