+++ fluxion-fluid/Cargo.toml
@@ -12,6 +12,7 @@ thiserror = "1.0"
 num-traits = "0.2"
 uom = { version = "0.35", features = ["f32"] }
 nalgebra = "0.34"
+uuid = { version = "1.0", features = ["serde"] }
 
 [dev-dependencies]
 approx = "0.5"
